clean: extract twemproxy options to resource_builders
