+++ pkg/resource_builders/twemproxy/options.go
@@ -1,4 +1,4 @@
-package config
+package twemproxy
 
 import (
 	"fmt"
